pyproject.toml
[tool.poetry]
name = "api-dashboard"
version = "0.3.2"
description = ""
authors = ["Ernesto Martinez del Pino <ernestomar1997@hotmail.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "0.*.*"
python-decouple = "^3.6"
uvicorn = "^0.18.3"
radar-mongodb = "0.3.*"

[tool.poetry.dev-dependencies]
coverage = {extras = ["toml"], version = "^6.3.2"}
flake8 = "^4.0.1"
flake8-bugbear = "^22.3.23"
pytest = "^7.1.1"
pytest-html = "^3.1.1"
pytest-cov = ">=2.10.1,<3.0.0"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [".*", "*/site-packages/*"]

[tool.coverage.report]
fail_under = 0

[[tool.poetry.source]]
name = "radar"
url = "https://pypiserver.radaroptions.io"
secondary = true